+++ scripts/fux-dice-roller.js
@@ -260,3 +260,5 @@ Hooks.on("renderSidebarTab", async (app, html) => {
 
 
 
+
+